Create a new hydraulics library for QElectroTech
The out-of-the-box library was missing a lot of components I wanted for
hydraulic diagrams.  As well, the quality of the existing components
wasn't as high as I'd like.  So I've added new components and modified
the old ones to make my own library.

The license is GPLv2 since it's required by QElectroTech' license.